fixed double checker, moved polygon functions to Geometry and separated functors
